+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2024-08-09
+
+### Added
+
+- Initial release of the Python script:
+  - Generates a Task Scheduler XML file
+  - Allows users to choose between Miniconda, Anaconda, and Local Python
+  - Prompts users to input the name of their environment (if using Miniconda or Anaconda)
+  - Uses the script directory as the output path
+  - Default trigger time is set to 6:00 AM on a specified day, with instructions to modify settings after importing to Task Scheduler
+  - Includes a batch file (`MS_reward.bat`) for automatic execution of the Python script
+
 ## [0.1.0] - 2024-07-27
 
 ### Added